+++ src/components/manage/Tiles/Video/Edit.jsx
@@ -1,29 +1,23 @@
 /**
- * Edit image tile.
- * @module components/manage/Tiles/Image/Edit
+ * Edit video tile.
+ * @module components/manage/Tiles/Title/Edit
  */
 
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
-import { injectIntl } from 'react-intl';
-import { readAsDataURL } from 'promise-file-reader';
-import { Button, Icon, Image, Message } from 'semantic-ui-react';
-import { bindActionCreators } from 'redux';
+import { defineMessages, injectIntl, intlShape } from 'react-intl';
+import { Button, Form, Input, Icon, Embed, Message } from 'semantic-ui-react';
 
-import { createContent } from '../../../../actions';
-import { getBaseUrl } from '../../../../helpers';
+const messages = defineMessages({
+  save: {
+    id: 'Save',
+    defaultMessage: 'Save',
+  },
+});
 
 @injectIntl
-@connect(
-  state => ({
-    request: state.content.create,
-    content: state.content.data,
-  }),
-  dispatch => bindActionCreators({ createContent }, dispatch),
-)
 /**
- * Edit image tile class.
+ * Edit video tile class.
  * @class Edit
  * @extends Component
  */
@@ -37,16 +31,10 @@ export default class Edit extends Component {
     selected: PropTypes.bool.isRequired,
     tile: PropTypes.string.isRequired,
     data: PropTypes.objectOf(PropTypes.any).isRequired,
-    content: PropTypes.objectOf(PropTypes.any).isRequired,
-    request: PropTypes.shape({
-      loading: PropTypes.bool,
-      loaded: PropTypes.bool,
-    }).isRequired,
-    pathname: PropTypes.string.isRequired,
     onChangeTile: PropTypes.func.isRequired,
     onSelectTile: PropTypes.func.isRequired,
     onDeleteTile: PropTypes.func.isRequired,
-    createContent: PropTypes.func.isRequired,
+    intl: intlShape.isRequired,
   };
 
   /**
@@ -58,55 +46,34 @@ export default class Edit extends Component {
   constructor(props) {
     super(props);
 
-    this.onUploadImage = this.onUploadImage.bind(this);
+    this.onChangeUrl = this.onChangeUrl.bind(this);
+    this.onSubmitUrl = this.onSubmitUrl.bind(this);
     this.state = {
-      uploading: false,
+      url: '',
     };
   }
 
   /**
-   * Component will receive props
-   * @method componentWillReceiveProps
-   * @param {Object} nextProps Next properties
+   * Change url handler
+   * @method onChangeUrl
+   * @param {Object} target Target object
    * @returns {undefined}
    */
-  componentWillReceiveProps(nextProps) {
-    if (
-      this.props.request.loading &&
-      nextProps.request.loaded &&
-      this.state.uploading
-    ) {
-      this.setState({
-        uploading: false,
-      });
-      this.props.onChangeTile(this.props.tile, {
-        ...this.props.data,
-        url: nextProps.content['@id'],
-      });
-    }
+  onChangeUrl({ target }) {
+    this.setState({
+      url: target.value,
+    });
   }
 
   /**
-   * Upload image handler
-   * @method onUploadImage
+   * Submit url handler
+   * @method onSubmitUrl
    * @returns {undefined}
    */
-  onUploadImage({ target }) {
-    const file = target.files[0];
-    readAsDataURL(file).then(data => {
-      const fields = data.match(/^data:(.*);(.*),(.*)$/);
-      this.setState({
-        uploading: true,
-      });
-      this.props.createContent(getBaseUrl(this.props.pathname), {
-        '@type': 'Image',
-        image: {
-          data: fields[3],
-          encoding: fields[2],
-          'content-type': fields[1],
-          filename: file.name,
-        },
-      });
+  onSubmitUrl() {
+    this.props.onChangeTile(this.props.tile, {
+      ...this.props.data,
+      url: this.state.url,
     });
   }
 
@@ -129,15 +96,16 @@ export default class Edit extends Component {
    * @returns {string} Markup for the component.
    */
   render() {
+    const { data } = this.props;
     return (
       <div
         onClick={() => this.props.onSelectTile(this.props.tile)}
         className={[
           'tile',
-          'image',
+          'video',
           'align',
           this.props.selected && 'selected',
-          this.props.data.align,
+          data.align,
         ]
           .filter(e => !!e)
           .join(' ')}
@@ -149,7 +117,7 @@ export default class Edit extends Component {
                 icon
                 basic
                 onClick={this.onAlignTile.bind(this, 'left')}
-                active={this.props.data.align === 'left'}
+                active={data.align === 'left'}
               >
                 <Icon name="align left" />
               </Button>
@@ -159,7 +127,7 @@ export default class Edit extends Component {
                 icon
                 basic
                 onClick={this.onAlignTile.bind(this, 'right')}
-                active={this.props.data.align === 'right'}
+                active={data.align === 'right'}
               >
                 <Icon name="align right" />
               </Button>
@@ -169,9 +137,7 @@ export default class Edit extends Component {
                 icon
                 basic
                 onClick={this.onAlignTile.bind(this, 'center')}
-                active={
-                  this.props.data.align === 'center' || !this.props.data.align
-                }
+                active={data.align === 'center' || !data.align}
               >
                 <Icon name="align center" />
               </Button>
@@ -181,7 +147,7 @@ export default class Edit extends Component {
                 icon
                 basic
                 onClick={this.onAlignTile.bind(this, 'full')}
-                active={this.props.data.align === 'full'}
+                active={data.align === 'full'}
               >
                 <Icon name="align justify" />
               </Button>
@@ -197,25 +163,54 @@ export default class Edit extends Component {
             </Button.Group>
           </div>
         )}
-        {this.props.data.url ? (
+        {data.url ? (
           <p>
-            <Image src={`${this.props.data.url}/@@images/image`} alt="" />
+            <div className="ui blocker" />
+            {data.url.match('list') ? (
+              <Embed
+                url={`https://www.youtube.com/embed/videoseries?list=${
+                  data.url.match(/^.*\?list=(.*)$/)[1]
+                }`}
+                icon="arrow right"
+                defaultActive
+                autoplay={false}
+              />
+            ) : (
+              <Embed
+                id={
+                  data.url.match(/.be\//)
+                    ? data.url.match(/^.*\.be\/(.*)/)[1]
+                    : data.url.match(/^.*\?v=(.*)$/)[1]
+                }
+                source="youtube"
+                icon="arrow right"
+                defaultActive
+                autoplay={false}
+              />
+            )}
           </p>
         ) : (
           <p>
             <Message>
               <center>
-                <h4>Image</h4>
-                <p>Upload a new image</p>
+                <h4>Video</h4>
+                <p>Specify a youtube video or playlist url</p>
                 <p>
-                  <label className="ui button file">
-                    Browse
-                    <input
-                      type="file"
-                      onChange={this.onUploadImage}
-                      style={{ display: 'none' }}
+                  <Form onSubmit={this.onSubmitUrl}>
+                    <Input onChange={this.onChangeUrl} />
+                    <Button
+                      basic
+                      circular
+                      primary
+                      icon="arrow right"
+                      type="submit"
+                      title={
+                        this.props.submitLabel
+                          ? this.props.submitLabel
+                          : this.props.intl.formatMessage(messages.save)
+                      }
                     />
-                  </label>
+                  </Form>
                 </p>
               </center>
             </Message>